move create_table_info into source::table::test_utils
