+++ src/batch/src/executor/insert.rs
@@ -161,7 +161,7 @@ mod tests {
     use risingwave_common::catalog::schema_test_utils;
     use risingwave_common::column_nonnull;
     use risingwave_common::types::DataType;
-    use risingwave_common::util::create_table_info;
+    use risingwave_source::table_test_utils::create_table_info;
     use risingwave_source::{MemSourceManager, SourceDescBuilder, SourceManagerRef};
     use risingwave_storage::memory::MemoryStateStore;
     use risingwave_storage::store::ReadOptions;
