+++ src/source/src/lib.rs
@@ -44,6 +44,7 @@ pub mod connector_source;
 pub mod monitor;
 pub mod row_id;
 mod table;
+pub use table::test_utils as table_test_utils;
 
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub enum SourceFormat {
